Add update-hash-table function to fix insertion order issue
While attempting to randomize the order in which I assoc keys into
the hash-map for testing, I discovered that the way that I was handling
updating the hash-table was based on a blind sort of the :key keyword
for the nodes in the map. The problem is, sometimes :sub-hash nodes
are in this map, so the ordering would then go haywire.

update-hash-table was introduced to ensure that the hash-tables are
sparse and properly arranged throughout operation, no matter what
happens.

I also discovered a bug when inserting 15 and 23, wrt to their
segment seqs.

15 -> [11 11 00 00]

23 -> [11 01 01 00]

Because the final 00 matched, the the part of the code in build-subhash
that constructs the :subhash node was running more than once, botching its
initial correct result.

I added unit tests to ensure that these problems that were discovered
remain fixed and documented.